Add files via upload
Tensor flow may update soon so the code may need an update. This code leverages the MNIST dataset for hand written digits, creates a generator model with 1 dense layer and 3 transposed 2D Convolution layers, each leveraging Batch normalization for faster training speeds and Leaky ReLu for activation. The code also creates a discriminator that consists of 2 Convolution layers that leverage Dropout in order to scale for over/underfitting, and Leaky Relu for activation. The discriminator also has a final Dense layer and is similar to a CNN Image Classifier. Both the Generator and Discriminator use Sigmoid cross entropy as loss functions and ADAM optimizers. The network is trained for 50 epochs and shows the progression for every epoch. In the end we create a Gif of the generated images.